Todd Isom coding challenge commits:
I would consider this not completely finished.  I would have liked to include more logging, error checking, along with adding caching for the list of all employees that is used often.

As for some of the existing problems, the delete employee endpoint should be based on id and not name.
The create employee endpoint should have a value for email instead of one being created. (probably).

Thanks for taking the time to review this.
Todd Isom